Add patch which includes cards with a menu for copying text (#1)
This patch allows results or links to be copied from from direct answers or help articles.

The patch produces the cards `allfields-standard-with-menu`, `documentsearch-standard-with-menu`, and `help-article-with-menu`. The `help-article-with-menu` is a fork of `document-standard`.

This patch also forks the iframe-common.js script and the custom-modules.js script. These files were based on Theme 1.24.

After applying the patch, the config must be updated to point to these new cards. Also, jambo.json needs to be updated to include `cards` and `directanswercards` in the partial dirs array.

Note: The menu button doesn't appear on the cards on the vertical page, only on the universal page. When it's enabled on verticals, the dropdown overflows the cards and gets cut off because we have `overflow: hidden` set on the first and last cards. This isn't an issue on universal on the template repo because the `overflow: hidden` was removed along with the card borders. Fixing this is out fo scope of this item.

J=SLAP-1682
TEST=manual

Build a new site on Theme 1.26 and apply the patch. Update the config to use the new cards and test out and confirm the copy buttons work.